Improvements to package structure, added new 'mesh' and 'utilities' modules
