+++ tests/test_geometry.py
@@ -6,8 +6,9 @@ from scipy.integrate import quad, simps
 
 from tokamesh.construction import equilateral_mesh
 from tokamesh import TriangularMesh
-from tokamesh.geometry import Camera, BarycentricGeometryMatrix, linear_geometry_matrix
-from tokamesh.geometry import radius_hyperbolic_integral, build_edge_map
+from tokamesh.geometry import BarycentricGeometryMatrix, linear_geometry_matrix
+from tokamesh.geometry import radius_hyperbolic_integral
+from tokamesh.utilities import build_edge_map, Camera
 
 
 def test_BarycentricGeometryMatrix():
